+++ Backend/src/Controllers/Users/tourGuide.controller.ts
@@ -30,7 +30,6 @@ export const createTourGuide = async (req: Request, res: Response) => {
 };
 export const getTourGuide = async (req: Request, res: Response) => {
 	try {
-		//maybe we need to add checker here based on the flow of the page
 		res.status(200).send(await TourGuide.find());
 	} catch (error) {
 		console.log(error);
@@ -39,6 +38,8 @@ export const getTourGuide = async (req: Request, res: Response) => {
 };
 export const updateTourGuide = async (req: Request, res: Response) => {
 	const id = req.params.id;
+	const tourGuide=await TourGuide.findById(id);
+	const Verified=tourGuide?.isVerified
 	const {
 		username,
 		email,
@@ -49,8 +50,8 @@ export const updateTourGuide = async (req: Request, res: Response) => {
 		previous,
 	} = req.body;
 	try {
-		//maybe we need to add checker here based on the flow of the page
-		const adv = await TourGuide.findByIdAndUpdate(
+		if(Verified){
+		res.status(200).send( await TourGuide.findByIdAndUpdate(
 			id,
 			{
 				username,
@@ -64,18 +65,27 @@ export const updateTourGuide = async (req: Request, res: Response) => {
 			{
 				new: true,
 			},
-		);
-		res.status(200).send(adv);
+		));
+	}
+	else{
+		res.status(500).send("user not Verified");
+	}
 	} catch (error) {
 		res.status(500).send("failed");
 	}
 };
 export const deleteTourGuide = async (req: Request, res: Response) => {
 	const id = req.params.id;
+	const tourGuide=await TourGuide.findById(id);
+	const Verified=tourGuide?.isVerified
 	try {
-		//maybe we need to add checker here based on the flow of the page
+		if(Verified){
 		await TourGuide.findByIdAndDelete(id);
 		res.status(200).send("tourGuide deleted successfully");
+		}
+		else{
+			res.status(500).send("user not Verified");
+		}
 	} catch (error) {
 		res.status(500).send("Failed to delete tourGuide");
 	}
@@ -83,8 +93,15 @@ export const deleteTourGuide = async (req: Request, res: Response) => {
 
 export const viewItinerary = async (req: Request, res: Response) => {
 	const id=req.params.id
+	const tourGuide=await TourGuide.findById(id);
+	const Verified=tourGuide?.isVerified
 	try {
-		res.status(200).send(await TourGuide.findById(id).select('itinerary'));
+		if(Verified){
+		res.status(200).send(await TourGuide.findById(id));
+		}
+		else{
+			res.status(500).send("user not Verified");
+		}
 	} catch (error) {
 		console.log(error);
 		res.status(500).send("failed");
